+++ pkg/providers/iam/provider.go
@@ -79,7 +79,9 @@ func (p *Subprovider) SDKDataSources() map[string]*schema.Resource {
 
 // FrameworkResources returns the IAM resources implemented using terraform-plugin-framework
 func (p *Subprovider) FrameworkResources() []func() resource.Resource {
-	return []func() resource.Resource{}
+	return []func() resource.Resource{
+		NewIPAllowlistResource,
+	}
 }
 
 // FrameworkDataSources returns the IAM data sources implemented using terraform-plugin-framework